refactor: change component file names to multi-word
To match `multi-word-component-names` eslint rule. Not a breaking change because they were all exported under these multi-word `Zoa` names anyway; this just changes the filenames.
