+++ src/components/index.js
@@ -1,8 +1,8 @@
-export { default as ZoaButton } from './buttons/button/Button.vue';
-export { default as ZoaToggleButton } from './buttons/togglebutton/ToggleButton.vue';
-export { default as ZoaTabs } from './buttons/tabs/Tabs.vue';
+export { default as ZoaButton } from './buttons/button/ZoaButton.vue';
+export { default as ZoaToggleButton } from './buttons/togglebutton/ZoaToggleButton.vue';
+export { default as ZoaTabs } from './buttons/tabs/ZoaTabs.vue';
 
-export { default as ZoaInput } from './inputs/Input.vue';
+export { default as ZoaInput } from './inputs/ZoaInput.vue';
 
-export { default as ZoaFlash } from './dialogs/flash/Flash.vue';
-export { default as ZoaModal } from './dialogs/modal/Modal.vue';
+export { default as ZoaFlash } from './dialogs/flash/ZoaFlash.vue';
+export { default as ZoaModal } from './dialogs/modal/ZoaModal.vue';
