+++ src/components/inputs/checkbox/ZoaCheckbox.stories.js
@@ -1,28 +1,28 @@
-import ZoaDateSimple from './DateSimple.vue';
+import ZoaCheckbox from './ZoaCheckbox.vue';
 import { ZoaInput } from '../../index.js';
 import { argTypes } from '../stories.js';
 import { renderSetup } from '../../utils/stories.js';
 
 const template = `
-<zoa-input zoa-type="date-simple"
+<zoa-input zoa-type="checkbox"
            :class="rootClass"
            :label="label"
            :label-position="labelPosition"
            :help="help"
            :help-position="helpPosition"
            :disabled="disabled"
-           :config="{delay, placeholder, min, max, step}"
+           :config="{delay, name, checkValue}"
 />
 `;
 
 const meta = {
-  component: ZoaDateSimple,
-  title: 'Components/Inputs/Date/Simple Date',
+  component: ZoaCheckbox,
+  title: 'Components/Inputs/Checkbox',
   argTypes,
   parameters: {
     docs: {
       description: {
-        component: 'A standard date picker.',
+        component: 'A single checkbox with an optional label.',
       },
       source: {
         code: template,
@@ -36,16 +36,14 @@ export default meta;
 const Base = {
   args: {
     class: '',
-    label: 'Date',
-    labelPosition: 'above',
+    label: 'Checkbox',
+    labelPosition: 'right',
     help: 'Some example help text.',
     helpPosition: 'right',
     disabled: false,
-    delay: 200,
-    placeholder: 'placeholder',
-    min: null,
-    max: null,
-    step: 'any',
+    delay: 0,
+    name: '',
+    checkValue: 'chkbx',
   },
   render: (args) => ({
     components: { ZoaInput },
