+++ src/components/inputs/date/ZoaDateAmbiguous.stories.js
@@ -1,35 +1,29 @@
-import ZoaRangeSlider from './RangeSlider.vue';
+import ZoaDateAmbiguous from './ZoaDateAmbiguous.vue';
 import { ZoaInput } from '../../index.js';
 import { argTypes } from '../stories.js';
 import { renderSetup } from '../../utils/stories.js';
 
 const template = `
-<zoa-input zoa-type="range-slider"
+<zoa-input zoa-type="date-ambiguous"
            :class="rootClass"
            :label="label"
            :label-position="labelPosition"
            :help="help"
            :help-position="helpPosition"
            :disabled="disabled"
-           :config="{delay, placeholder, min, max, step, labelsRight,
-                     labelUpper, labelLower}"
+           :config="{delay, placeholder, min, max}"
 />
 `;
 
 const meta = {
-  component: ZoaRangeSlider,
-  title: 'Components/Inputs/Slider/Range',
-  argTypes: {
-    ...argTypes,
-    labelsRight: {
-      control: 'boolean',
-    },
-  },
+  component: ZoaDateAmbiguous,
+  title: 'Components/Inputs/Date/Ambiguous Date',
+  argTypes,
   parameters: {
     docs: {
       description: {
         component:
-          'A component with two sliders representing a lower and upper value. Returns the values as an array.',
+          'A date picker that allows for missing parts, e.g. just a year, a year and a month, a month and a day, etc.',
       },
       source: {
         code: template,
@@ -42,19 +36,16 @@ export default meta;
 
 const Base = {
   args: {
-    label: 'Range',
+    class: '',
+    label: 'Date',
     labelPosition: 'above',
     help: 'Some example help text.',
     helpPosition: 'right',
     disabled: false,
     delay: 200,
-    placeholder: null,
-    min: 0,
-    max: 100,
-    step: 1,
-    labelsRight: false,
-    labelUpper: 'Upper',
-    labelLower: 'Lower',
+    placeholder: 'placeholder',
+    min: '0000-01-01',
+    max: 'today',
   },
   render: (args) => ({
     components: { ZoaInput },
